tests/requirements-test.txt
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
requests==2.32.3
httpx==0.25.2
python-dotenv==1.0.0
python-jose==3.3.0
